Revert "fix(chat): fix error when send publication request for chats which ha…"
This reverts commit 814705890fd07798e64adc5a734fb488a5d3fa57.
