+++ apps/chat/src/components/Chat/Publish/PublishDialog.tsx
@@ -16,7 +16,7 @@ import {
 } from '@/src/utils/app/id';
 import { EnumMapper } from '@/src/utils/app/mappers';
 import {
-  buildDedupedPublicationFileTargetsFromConversations,
+  createFoldersFilesTargetUrl,
   isEntityIdPublic,
 } from '@/src/utils/app/publications';
 import { NotReplayFilter } from '@/src/utils/app/search';
@@ -54,12 +54,46 @@ import { Modal } from '@/src/components/Common/Modal';
 import { CreatePublicationHandler } from './PublicationHandler/CreatePublicationHandler';
 
 import { Conversation, PublishActions } from '@epam/ai-dial-shared';
+import compact from 'lodash-es/compact';
+import escapeRegExp from 'lodash-es/escapeRegExp';
+import flatMapDeep from 'lodash-es/flatMapDeep';
 
 interface PublishDialogContainerProps {
   publicationModel: PublicationModel;
   resourceType: BackendResourceType;
   filteredConversationFiles: DialFile[];
 }
+
+const transformFoldersFilesIds = (
+  conversations: Conversation[],
+  entityFolderId: string,
+) => {
+  const folderOldPathPartsRegExp = new RegExp(
+    escapeRegExp(getIdWithoutRootPathSegments(entityFolderId)),
+  );
+
+  return conversations.flatMap((c) => {
+    const urls = compact(
+      flatMapDeep(c.playback?.messagesStack || c.messages, (m) =>
+        m.custom_content?.attachments?.map((a) => a.url),
+      ),
+    );
+
+    return urls.map((oldUrl) => {
+      const decodedOldUrl = ApiUtils.decodeApiUrl(oldUrl);
+
+      return {
+        oldUrl: decodedOldUrl,
+        newUrl: createFoldersFilesTargetUrl(
+          constructPath(
+            getFolderIdFromEntityId(c.id),
+            ...decodedOldUrl.split('/').slice(-1),
+          ).replace(folderOldPathPartsRegExp, ''),
+        ),
+      };
+    });
+  });
+};
 
 const PublishDialogContainer = ({
   publicationModel,
@@ -142,35 +176,32 @@ const PublishDialogContainer = ({
       };
     });
 
-    const mappedWithConversationsFiles =
-      buildDedupedPublicationFileTargetsFromConversations(
-        filteredEntities as Conversation[],
-        entity.folderId,
-        isFolder,
-      );
+    const mappedWithConversationsFiles = transformFoldersFilesIds(
+      filteredEntities as Conversation[],
+      entity.folderId,
+    );
 
     const fileResources =
       action === PublishActions.DELETE
         ? []
-        : isFolder
-          ? mappedWithConversationsFiles.map(({ oldUrl, newUrl }) => ({
+        : filteredConversationFiles.map(({ id }) => {
+            const decodedId = ApiUtils.decodeApiUrl(id);
+
+            const url =
+              (isFolder
+                ? mappedWithConversationsFiles.find(
+                    (file) => file.oldUrl === decodedId,
+                  )?.newUrl
+                : transformIdToRootEntityId(decodedId)) ??
+              transformIdToRootEntityId(decodedId);
+
+            return {
               action: PublishActions.ADD_IF_ABSENT,
-              sourceUrl: oldUrl,
-              reviewUrl: oldUrl,
-              targetUrl: replaceIdWithBucket(newUrl, PUBLIC_URL_PREFIX),
-            }))
-          : filteredConversationFiles.map(({ id }) => {
-              const decodedId = ApiUtils.decodeApiUrl(id);
-
-              const url = transformIdToRootEntityId(decodedId);
-
-              return {
-                action: PublishActions.ADD_IF_ABSENT,
-                sourceUrl: decodedId,
-                reviewUrl: decodedId,
-                targetUrl: replaceIdWithBucket(url, PUBLIC_URL_PREFIX),
-              };
-            });
+              sourceUrl: decodedId,
+              reviewUrl: decodedId,
+              targetUrl: replaceIdWithBucket(url, PUBLIC_URL_PREFIX),
+            };
+          });
 
     const iconResource = [];
     if (
