--- apps/chat/src/utils/app/__tests__/publications.test.ts
@@ -1,89 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-import { transformIdToRootEntityId } from '@/src/utils/app/id';
-import { buildDedupedPublicationFileTargetsFromConversations } from '@/src/utils/app/publications';
-import { ApiUtils } from '@/src/utils/server/api';
-
-import { Conversation } from '@epam/ai-dial-shared';
-
-describe('buildDedupedPublicationFileTargetsFromConversations', () => {
-  const entityFolderId = 'conversations/mybucket/folder02';
-  const sharedFile = 'files/mybucket/docs/shared.pdf';
-
-  const convNested = {
-    id: 'conversations/mybucket/folder02/nested/chat-1',
-    messages: [
-      {
-        custom_content: {
-          attachments: [{ url: sharedFile }],
-        },
-      },
-    ],
-  } as Conversation;
-
-  const convRoot = {
-    id: 'conversations/mybucket/folder02/chat-2',
-    messages: [
-      {
-        custom_content: {
-          attachments: [{ url: sharedFile }],
-        },
-      },
-    ],
-  } as Conversation;
-
-  it('returns one target when the same file is attached under two chats in a folder publish', () => {
-    const result = buildDedupedPublicationFileTargetsFromConversations(
-      [convNested, convRoot],
-      entityFolderId,
-      true,
-    );
-
-    expect(result).toHaveLength(1);
-    expect(result[0].oldUrl).toBe(ApiUtils.decodeApiUrl(sharedFile));
-
-    const fromRoot = buildDedupedPublicationFileTargetsFromConversations(
-      [convRoot],
-      entityFolderId,
-      true,
-    );
-    expect(result[0].newUrl).toBe(fromRoot[0].newUrl);
-  });
-
-  it('uses the shortest target path regardless of conversation order', () => {
-    const forward = buildDedupedPublicationFileTargetsFromConversations(
-      [convNested, convRoot],
-      entityFolderId,
-      true,
-    );
-    const reversed = buildDedupedPublicationFileTargetsFromConversations(
-      [convRoot, convNested],
-      entityFolderId,
-      true,
-    );
-
-    expect(forward).toHaveLength(1);
-    expect(reversed).toHaveLength(1);
-    expect(reversed[0].newUrl).toBe(forward[0].newUrl);
-
-    const fromRoot = buildDedupedPublicationFileTargetsFromConversations(
-      [convRoot],
-      entityFolderId,
-      true,
-    );
-    expect(forward[0].newUrl).toBe(fromRoot[0].newUrl);
-    expect(reversed[0].newUrl).toBe(fromRoot[0].newUrl);
-  });
-
-  it('uses transformIdToRootEntityId when not publishing a folder', () => {
-    const decoded = ApiUtils.decodeApiUrl(sharedFile);
-    const result = buildDedupedPublicationFileTargetsFromConversations(
-      [convNested],
-      entityFolderId,
-      false,
-    );
-
-    expect(result).toHaveLength(1);
-    expect(result[0].newUrl).toBe(transformIdToRootEntityId(decoded));
-  });
-});
